advanced-analytics/recommendation-engine/requirements.txt
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
google-api-core==1.27.0
google-auth==1.30.0
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
joblib==1.0.1
numpy==1.20.3
packaging==20.9
pandas==1.2.4
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pyportfolioopt==1.4.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.26.4
Flask==2.0.0
gunicorn==20.1.0  # https://github.com/benoitc/gunicorn
yfinance==0.1.59
fsspec==2021.5.0
gcsfs==2021.5.0
